Fixed Radiobuttons and Languagesetup

+++ Kext Updater/scripts/nightly.command
@@ -1,4 +1,6 @@
 #!/bin/bash
 
 
+if [ -f /tmp/nightly.tmp ]; then
 rm /tmp/nightly.tmp
+fi
